+++ kernel/main.go
@@ -6,6 +6,7 @@ import (
 
 	"publish/client"
 	"publish/config"
+	"publish/models"
 	"publish/server"
 )