+++ kernel/models/interface.go
@@ -1,0 +1,6 @@
+package models
+
+type model interface {
+	Empty() bool
+	One(id string) interface{}
+}