+++ kernel/server/middleware/auth/access/dynamic.go
@@ -12,37 +12,24 @@ import (
 
 func Dynamic(c *gin.Context) (bool, func(c *gin.Context), func(c *gin.Context)) {
 	var (
-		r       *client.ResponseBody
-		err     error
-		err_msg string
-		data    []interface{}
-		path    string
-		docs    []string
+		r    *client.ResponseBody
+		err  error
+		data []interface{}
+		path string
+		docs []string
 	)
-	/* 通过 API 查询具有访问控制字段的文档 ID 列表 */
-	r, err = client.GetDocsByAttrName(client.C.R(), config.C.Siyuan.Publish.Access.Name)
-	r, err_msg = client.Response(r, err)
-	if r == nil {
+
+	/* 获得访问控制列表 root_id => "public" | "protected" | "private" */
+	acl, err_msg := client.GetACL()
+	if acl == nil {
 		return false, func(c *gin.Context) { status.S.StatusInternalServerError(c, err_msg) }, nil
 	}
-
-	/* 建立文档 ID => 访问权限类型 public/protected/private 的映射 */
-	data = r.Data.([]interface{})
-	docs_with_access := make(map[string]string)
-	switch {
-	case len(data) == 0:
+	if len(acl) == 0 {
 		return false, status.S.StatusAccessDenied, nil
-	default:
-		for _, v := range data {
-			docs_with_access[v.(map[string]interface{})["root_id"].(string)] = v.(map[string]interface{})["value"].(string)
-		}
 	}
 
 	/* 获得文档路径 */
 	id := c.GetString("id")
-	if id == "" {
-		return false, status.S.StatusParamsError, nil
-	}
 	r, err = client.GetBlockByID(client.C.R(), id)
 	r, err_msg = client.Response(r, err)
 	if r == nil {
@@ -63,9 +50,9 @@ func Dynamic(c *gin.Context) (bool, func(c *gin.Context), func(c *gin.Context)) 
 	for i, j := 0, len(docs)-1; i < j; i, j = i+1, j-1 {
 		docs[i], docs[j] = docs[j], docs[i]
 	}
-	// fmt.Printf("%#v\n%#v\n", docs_with_access, docs)
+	// fmt.Printf("%#v\n%#v\n", acl, docs)
 	for _, doc_id := range docs {
-		switch docs_with_access[doc_id] {
+		switch acl[doc_id] {
 		case config.C.Siyuan.Publish.Access.Public.Value:
 			c.Set("access", config.C.Siyuan.Publish.Access.Public.Value)
 			return true, nil, nil
@@ -74,6 +61,8 @@ func Dynamic(c *gin.Context) (bool, func(c *gin.Context), func(c *gin.Context)) 
 			// TODO
 			fallthrough
 		case config.C.Siyuan.Publish.Access.Private.Value:
+			fallthrough
+		default:
 			c.Set("access", config.C.Siyuan.Publish.Access.Private.Value)
 			return false, status.S.StatusAccessDenied, nil
 		}